[aievec] Move exp lowering pattern to vector-2-aievec
The lowering pattern and legalization of math::ExpOp was in the wrong
pipeline, this patch moves it to the Vec2AIEVec pipeline.
