+++ lib/Dialect/AIEVec/Transforms/VectorToAIEVecConversions.cpp
@@ -24,6 +24,7 @@
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/EmitC/IR/EmitC.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/Math/IR/Math.h"
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/IR/PatternMatch.h"
@@ -1683,6 +1684,45 @@ struct FuseExtIntoUPDPattern : public OpConversionPattern<aievec::ExtOp> {
   }
 };
 
+// Lower ExpOp to function call
+struct ComputeExpOpByLUTPattern : public OpConversionPattern<math::ExpOp> {
+  using OpConversionPattern<math::ExpOp>::OpConversionPattern;
+
+  LogicalResult
+  matchAndRewrite(math::ExpOp expOp, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    VectorType srcType = dyn_cast<VectorType>(adaptor.getOperand().getType());
+
+    if (!srcType) {
+      return failure();
+    }
+
+    Type scalarType = srcType.getElementType();
+    unsigned elWidth = scalarType.getIntOrFloatBitWidth();
+    unsigned laneSize = getVectorLaneSize(srcType);
+    if (!isa<FloatType>(scalarType) || laneSize != 16 || elWidth != 16)
+      return failure();
+
+    StringRef includeName = "exp_lut.h";
+    ModuleOp moduleOp = expOp->getParentOfType<mlir::ModuleOp>();
+    rewriter.setInsertionPointToStart(
+        &moduleOp.getRegion().getBlocks().front());
+    rewriter.create<emitc::IncludeOp>(moduleOp.getLoc(), includeName, false);
+
+    SmallVector<Value> expOperands = {adaptor.getOperand()};
+
+    rewriter.setInsertionPoint(expOp);
+    Type accType = getVectorOpDestType(srcType, /*AIEML =*/true);
+    auto funcOp = rewriter.create<emitc::CallOp>(
+        expOp.getLoc(), TypeRange{accType}, "getExpBf16", nullptr, nullptr,
+        expOperands);
+    rewriter.replaceOpWithNewOp<aievec::SRSOp>(expOp, srcType,
+                                               funcOp.getResult(0));
+
+    return success();
+  }
+};
+
 //===----------------------------------------------------------------------===//
 // Pattern collection
 //===----------------------------------------------------------------------===//
@@ -1712,6 +1752,7 @@ static void populateAIEVecV2ConversionPatterns(RewritePatternSet &patterns,
   patterns.add<
       LowerVectorAddIOpToAIEVecAddElemOp,
       LowerVectorSubIOpToAIEVecSubElemOp,
+      ComputeExpOpByLUTPattern,
       ConvertMulIToAIEVecMulElemOpPattern,
       LowerVectorAddFOpToAIEVecAddElemOp,
       LowerVectorSubFOpToAIEVecSubElemOp,
@@ -1741,9 +1782,22 @@ static void populateAIEVecV2ConversionPatterns(RewritePatternSet &patterns,
 // TODO: Review the validity of these legalizations beyond basic cases.
 static void configureAIEVecCommonLegalizations(ConversionTarget &target,
                                                AnalysisManager &am) {
-  target.addLegalDialect<xilinx::aievec::AIEVecDialect, arith::ArithDialect>();
+  target.addLegalDialect<xilinx::aievec::AIEVecDialect, arith::ArithDialect,
+                         emitc::EmitCDialect>();
   target.addIllegalOp<vector::TransferReadOp>();
   target.addIllegalOp<vector::ExtractStridedSliceOp>();
+  target.addDynamicallyLegalOp<math::ExpOp>([](math::ExpOp expOp) {
+    VectorType srcType = dyn_cast<VectorType>(expOp.getOperand().getType());
+    if (!srcType) {
+      return true;
+    }
+    Type scalarType = srcType.getElementType();
+    unsigned elWidth = scalarType.getIntOrFloatBitWidth();
+    unsigned laneSize = getVectorLaneSize(srcType);
+    if (!isa<FloatType>(scalarType) || laneSize != 16 || elWidth != 16)
+      return true;
+    return false;
+  });
   target.addDynamicallyLegalOp<arith::AddIOp>(
       [](arith::AddIOp op) { return !isa<VectorType>(op.getType()); });
   target.addDynamicallyLegalOp<arith::AddFOp>(
